pyproject.toml
[tool.poetry]
name = "djangoexamples"
version = "0.1.0"
description = "Source code for Django related videos"
authors = ["Yuri Mednikov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
